requirements.txt
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dialogflow==1.1.0
Flask==1.1.2
google-api-core==1.23.0
google-auth==1.23.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pytz==2020.4
requests==2.25.0
rsa==4.6
six==1.15.0
twilio==6.48.0
urllib3==1.26.2
Werkzeug==1.0.1
